Move Unix Icon into regular folder structure.
Also, remove its ISerializable implementation, as required.
