+++ src/System.Drawing.Common/src/System/Drawing/Icon.Serializable.Windows.cs
