+++ src/System.Drawing.Common/src/System/Drawing/Icon.Windows.cs
